test: move managed automation scenarios to integration lane
